requirement.txt
asgiref==3.7.2
Django~=5.0.4
Pillow==10.3.0
sqlparse==0.5.0
django-braces==1.15.0
django-embed-video==1.4.9
pymemcache==4.0.0
django-debug-toolbar==4.3.0
django-redisboard==8.4.0
djangorestframework==3.15.1
requests==2.31.0